Split kitplayer.c to multiple files
