+++ include/kitchensink/kitchensink.h
@@ -3,6 +3,7 @@
 
 #include "kitchensink/kitlib.h"
 #include "kitchensink/kiterror.h"
+#include "kitchensink/kitformats.h"
 #include "kitchensink/kitsource.h"
 #include "kitchensink/kitplayer.h"
 #include "kitchensink/kitutils.h"
